fix AMD export
use a bundler to generate a valid UMD build to prevent
a fatal error when the module is loaded by Webpack/Angular
etc.:

>  Super expression must either be null or a function

+ add an ESM build
