+++ .gitignore
@@ -2,3 +2,5 @@
 node_modules
 *.log
 *.swp
+/emitter.*
+/dist
