+++ .jshintrc
@@ -1,4 +1,6 @@
 {
+  "esversion": 6,
+
   "camelcase": true,
   "eqeqeq": true,
   "immed": true,
