+++ .travis.yml
@@ -1,9 +1,9 @@
 language: node_js
-
+cache:
+  yarn: true
+  directories:
+    - node_modules
 node_js:
-  - "0.10"
-  - "0.12"
-  - iojs
-
-script:
-  - npm test
+  - '8'
+  - '9'
+  - '10'
