+++ package.json
@@ -1,15 +1,30 @@
 {
   "name": "little-emitter",
   "version": "0.1.3",
-  "description": "A tiny event emitter for node and browser",
-  "main": "emitter.js",
+  "description": "A tiny event emitter for node and the browser",
+  "main": "dist/emitter.js",
+  "module": "dist/emitter.mjs",
+  "umd": "dist/emitter.js",
+  "umd:main": "dist/emitter.js",
+  "source": "src/emitter.js",
   "author": {
     "name": "Alex Chao",
     "email": "alexchao1990@gmail.com"
   },
   "scripts": {
-    "test": "gulp test"
+    "build": "npm run build:esm && npm run build:umd && npm run build:umd:min",
+    "build:esm": "microbundle --no-sourcemap --strict --format es --no-compress",
+    "build:umd": "microbundle --no-sourcemap --strict --format umd --name Emitter --no-compress",
+    "build:umd:min": "microbundle --no-sourcemap --strict --format umd --name Emitter --output dist/emitter.min.js",
+    "lint": "jshint ./src",
+    "prepublishOnly": "npm run test",
+    "test": "npm run build && npm run lint && mocha ./test"
   },
+  "files": [
+    "dist/emitter.js",
+    "dist/emitter.min.js",
+    "dist/emitter.mjs"
+  ],
   "repository": {
     "type": "git",
     "url": "git://github.com/Alex1990/little-emitter.git"
@@ -25,17 +40,14 @@
   ],
   "dependencies": {},
   "devDependencies": {
-    "gulp": "^3.8.11",
-    "gulp-jshint": "^1.9.4",
-    "gulp-load-plugins": "^0.8.1",
-    "gulp-mocha": "^2.0.1",
-    "gulp-rename": "^1.2.0",
-    "gulp-uglify": "^1.1.0",
-    "mocha": "^2.2.4",
-    "should": "^6.0.1"
+    "jshint": "^2.9.7",
+    "microbundle": "^0.9.0",
+    "mocha": "^5.2.0",
+    "should": "^13.2.3"
   },
   "keywords": [
     "event",
-    "emitter"
+    "emitter",
+    "eventemitter"
   ]
 }
